Fixed the image size for texture scrambling so that it handles 1024 x 768 for the MRI
Now reads in and time stamps '5' keystrokes in the mri mode. It will abort a blank part of the ISI when it gets a '5' key essentially triggering at the pulse (then does stationary ISI and the trial). So for the MRI study the config file could have a really long blank ISI that essentially sets up for a MRI pulse trigger.